backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
authlib==1.2.0
httpx==0.23.3
python-dotenv==1.0.0
sqlite3==3.36.0
pytest==7.3.1
openai==0.27.0
boto3==1.26.155